refactor: move lststack tests to subdirectory
