+++ hera_cal/lst_stack/tests/__init__.py
@@ -1,0 +1,6 @@
+import pytest
+
+pytestmark = pytest.mark.filterwarnings(
+    "ignore:The uvw_array does not match the expected values given the antenna positions.",
+    "ignore:.*Using known values for HERA",
+)
